chore(test): refactored hash method into util
moved the hashing util out of the business
logic and added test coverage.

Fixes N/A
