+++ lib/karma-webpack.js
@@ -4,6 +4,8 @@ const fs = require('fs');
 const glob = require('glob');
 const minimatch = require('minimatch');
 
+const { hash } = require('./utils/hash');
+
 const { KarmaWebpackController } = require('./KarmaWebpackController');
 
 const controller = new KarmaWebpackController();
@@ -35,18 +37,6 @@ function registerExtraWebpackFiles(config, _controller) {
     served: true,
     watched: false,
   });
-}
-
-/**
- * Simple hash function by bryc
- * https://gist.github.com/iperelivskiy/4110988#gistcomment-2697447
- */
-function hash(s) {
-  let h = 0xdeadbeef;
-  for (let i = 0; i < s.length; i++) {
-    h = Math.imul(h ^ s.charCodeAt(i), 2654435761); // eslint-disable-line no-bitwise
-  }
-  return (h ^ (h >>> 16)) >>> 0; // eslint-disable-line no-bitwise
 }
 
 function getPathKey(filePath, withExtension = false) {
